CODETOOLS-7903326: Refactor CharacterRangeTableAttribute
