+++ src/classes/com/sun/tdk/jcov/instrument/asm/BranchCodeMethodAdapter.java
@@ -29,6 +29,7 @@ import java.util.*;
 
 import static org.objectweb.asm.Opcodes.*;
 
+import com.sun.tdk.jcov.instrument.CharacterRangeTable;
 import com.sun.tdk.jcov.instrument.Constants;
 import com.sun.tdk.jcov.instrument.DataBlock;
 import com.sun.tdk.jcov.instrument.DataBlockCatch;
@@ -355,10 +356,10 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                     continue;
                 }
 
-                for (CharacterRangeTableAttribute.CRTEntry entry : method().getCharacterRangeTable().getEntries()) {
+                for (CharacterRangeTable.CRTEntry entry : method().getCharacterRangeTable().getEntries()) {
                     if (entry.startBCI() == bci) {
 
-                        if ((entry.flags & CharacterRangeTableAttribute.CRTEntry.CRT_STATEMENT) != 0 /*& newBlock*/) {
+                        if ((entry.flags & CharacterRangeTable.CRTEntry.CRT_STATEMENT) != 0 /*& newBlock*/) {
                             newBlock = false;
                             if (insnToBB.get(insn) == null) {
                                 //System.out.println("Should add block at: " + bci + " in " + method().name +
@@ -368,7 +369,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
                             }
                         }
                     } else {
-                        if (entry.endBCI() == index && (entry.flags & CharacterRangeTableAttribute.CRTEntry.CRT_FLOW_TARGET) != 0) {
+                        if (entry.endBCI() == index && (entry.flags & CharacterRangeTable.CRTEntry.CRT_FLOW_TARGET) != 0) {
                             newBlock = true;
                         }
                     }
@@ -510,7 +511,7 @@ class BranchCodeMethodAdapter extends OffsetRecordingMethodAdapter {
     public void visitAttribute(Attribute attr) {
         super.visitAttribute(attr);
         if (attr instanceof CharacterRangeTableAttribute) {
-            method().setCharacterRangeTable((CharacterRangeTableAttribute) attr);
+            method().setCharacterRangeTable(((CharacterRangeTableAttribute) attr).crt);
         }
     }
